requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pytest==9.0.2
pytest-mock==3.15.1
httpx==0.25.2
debugpy==1.6.6
PyJWT==2.8.0
requests==2.31.0
